+++ client/src/components/ProductCard/ProductsPage.css
@@ -1,0 +1,14 @@
+.products {
+  text-align: right;
+}
+
+.icon {
+  border-radius: 50%;
+  margin-left: 10px;
+  margin-right: 10px;
+}
+
+.header {
+  display: flex;
+  justify-content: flex-end;
+}
