--- client/src/pages/products/Card.css
@@ -1,49 +1,0 @@
-.card {
-  /* From https://css.glass */
-  background: rgba(255, 255, 255, 0);
-  border-radius: 16px;
-  box-shadow: 0 4px 30px rgba(0, 0, 0, 0.1);
-  backdrop-filter: blur(0px);
-  -webkit-backdrop-filter: blur(0px);
-  width: 20rem;
-  /*height: 4rem; */
-  padding: 10px;
-  border: 1px solid rgba(255, 255, 255, 0.9);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-button {
-  width: 50px;
-  /* height: "50px"; */
-  border-radius: "50%";
-}
-
-h2 {
-  font-size: 1.5rem;
-  font-weight: bold;
-  color: black;
-}
-
-img {
-  width: 150px;
-}
-
-.buttons {
-  display: flex;
-  flex-direction: row-reverse;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.card-content {
-  display: flex;
-  flex-direction: row-reverse;
-  justify-content: space-around;
-  align-items: center;
-  width: 100%;
-}
-
-span {
-  margin: 10px;
-}
