--- client/src/pages/products/ProductsPage.css
@@ -1,14 +1,0 @@
-.products {
-  text-align: right;
-}
-
-.icon {
-  border-radius: 50%;
-  margin-left: 10px;
-  margin-right: 10px;
-}
-
-.header {
-  display: flex;
-  justify-content: flex-end;
-}
