+++ package.json
@@ -2,5 +2,8 @@
   "dependencies": {
     "react": "^18.3.1",
     "react-router-dom": "^6.23.1"
+  },
+  "devDependencies": {
+    "@types/node": "^20.14.10"
   }
 }
